requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.6
django-choices==1.7.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
